% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolleigen.R
\name{roll_pcr}
\alias{roll_pcr}
\title{Rolling Principal Component Regressions}
\usage{
roll_pcr(x, y, width, n_comps = ncol(x), weights = rep(1, width),
  intercept = TRUE, center = TRUE, scale = FALSE, min_obs = width,
  complete_obs = TRUE, na_restore = FALSE, online = TRUE)
}
\arguments{
\item{x}{vector or matrix. Rows are observations and columns are the independent variables.}

\item{y}{vector or matrix. Rows are observations and columns are the dependent variables.}

\item{width}{integer. Window size.}

\item{n_comps}{integer. Number of principal components.}

\item{weights}{vector. Weights for each observation within a window.}

\item{intercept}{logical. Either \code{TRUE} to include or \code{FALSE} to remove the intercept.}

\item{center}{logical. If \code{TRUE} then the weighted mean of each variable is used,
if \code{FALSE} then zero is used.}

\item{scale}{logical. If \code{TRUE} then the weighted standard deviation of each variable is used,
if \code{FALSE} then no scaling is done.}

\item{min_obs}{integer. Minimum number of observations required to have a value within a window,
otherwise result is \code{NA}.}

\item{complete_obs}{logical. If \code{TRUE} then rows containing any missing values are removed,
if \code{FALSE} then pairwise is used.}

\item{na_restore}{logical. Should missing values be restored?}

\item{online}{logical. Process observations using an online algorithm.}
}
\value{
A list containing the following components:
\item{coefficients}{A list of objects with the rolling and expanding coefficients for each \code{y}.
An object is the same class and dimension (with an added column for the intercept) as \code{x}.}
\item{r.squared}{A list of objects with the rolling and expanding r-squareds for each \code{y}.
An object is the same class as \code{x}.}
}
\description{
A function for computing the rolling and expanding principal component regressions of time-series data.
}
\examples{
n <- 15
m <- 3
x <- matrix(rnorm(n * m), nrow = n, ncol = m)
y <- rnorm(n)
weights <- 0.9 ^ (n:1)

# rolling regressions with complete windows
roll_pcr(x, y, width = 5, n_comps = 1)

# rolling regressions with partial windows
roll_pcr(x, y, width = 5, n_comps = 1, min_obs = 1)

# expanding regressions with partial windows
roll_pcr(x, y, width = n, n_comps = 1, min_obs = 1)

# expanding regressions with partial windows and weights
roll_pcr(x, y, width = n, n_comps = 1, min_obs = 1, weights = weights)
}
