\name{rolleigen-package}
\alias{rolleigen-package}
\alias{rolleigen}
\docType{package}
\title{
  \packageTitle{rolleigen}
}
\description{
  \packageDescription{rolleigen}
}
\details{
  \code{rolleigen} is a package that provides fast and efficient computation of rolling and expanding eigenanalysis for time-series data.
}
\author{
 \packageAuthor{rolleigen}
}
\references{
  Hotelling, H. (1933). "Analysis of a Complex of Statistical Variables Into Principal Components." \emph{Journal of Educational Psychology}, 24(6), 417-441. \doi{10.1037/h0071325}

  Massy, W.F. (1965). "Principal Components Regression in Exploratory Statistical Research." \emph{Journal of the American Statistical Association}, 60(309), 234-256. \doi{10.1080/01621459.1965.10480787}
  
  Pearson, K. (1901). "On Lines and Planes of Closest Fit to Systems of Points in Space." \emph{Philosophical Magazine}, 2(11), 559-572. \doi{10.1080/14786440109462720}
}
