\name{rollpca-package}
\alias{rollpca-package}
\alias{rollpca}
\docType{package}
\title{
  \packageTitle{rollpca}
}
\description{
  \packageDescription{rollpca}
}
\details{
  \code{rollpca} is a package that provides fast and efficient computation of rolling and expanding principal component analysis for time-series data.
}
\author{
 \packageAuthor{rollpca}
}