\name{rolleigen-package}
\alias{rolleigen-package}
\alias{rolleigen}
\docType{package}
\title{
  \packageTitle{rolleigen}
}
\description{
  \packageDescription{rolleigen}
}
\details{
  \code{rolleigen} is a package that provides fast and efficient computation of rolling and expanding eigenanalysis for time-series data.
}
\author{
 \packageAuthor{rolleigen}
}